+++ changelog.md
@@ -33,6 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * `GET    /api/roles/[ROLE_ID]/users`, to get all users associated to the role.
     * `POST   /api/roles/[ROLE_ID]/users`, to associate role and user.
     * `DELETE /api/roles/[ROLE_ID]/users/[USER_LOGIN]`, to dissociate role and user.
+    * `GET    /api/roles/[ROLE_ID]/groups`, to get all groups associated to the role.
+    * `POST   /api/roles/[ROLE_ID]/groups`, to associate role and group.
+    * `DELETE /api/roles/[ROLE_ID]/groups/[GROUP_ID]`, to dissociate role and group.
   * For all groups:
     * `GET    /api/groups`, to get all groups.
     * `POST   /api/groups`, to create a group.
@@ -45,5 +48,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * `GET    /api/groups/[GROUP_ID]/groups`, to get all groups associated to the group.
     * `POST   /api/groups/[GROUP_ID]/groups`, to associate group and group.
     * `DELETE /api/groups/[GROUP_ID]/groups/[GROUP_ID]`, to dissociate group and group.
+    * `GET    /api/groups/[GROUP_ID]/roles`, to get all roles associated to the group.
   * `/api/login`, to login.
   * `/api/redirect`, to redirect with token on leto-modelizer/leto-modelizer-admin.